+++ var/spack/repos/builtin/packages/libnbc/package.py
