+++ var/spack/repos/builtin/packages/r-r6/package.py
